refactor: rename `/utils` directory to `/utilities`
