+++ src/components/InclusionsTableBody.tsx
@@ -3,7 +3,7 @@ import { PayloadActionCreator } from "typesafe-actions";
 
 import { BaseEntityModel, TableViewModel } from "~/typeDefs";
 
-import { booleanToYesNo } from "~/utils";
+import { booleanToYesNo } from "~/utilities";
 
 import InclusionsTableCheckboxCell from "./InclusionsTableCheckboxCell";
 import InclusionsTableRow from "./InclusionsTableRow";
