+++ src/index.tsx
@@ -6,8 +6,8 @@ import { Provider } from "react-redux";
 
 import { configureStore } from "./redux/configureStore";
 import Routes from "./Routes";
-import { initInterceptor } from "./utils/httpInterceptor";
-import { initAnalytics } from "./utils/initAnalytics";
+import { initInterceptor } from "./utilities/httpInterceptor";
+import { initAnalytics } from "./utilities/initAnalytics";
 
 const history = createBrowserHistory();
 const store = configureStore(history);
