+++ src/utilities/initAnalytics.ts
